+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [unreleased]
 
+- Add: Added `ostrich.jumpToSlice()`
 - Add: Implemented basic rendering
 
 ## 0.1.0